Introduce composable for common functions needed in nitro and module and CWA plugin
