+++ src/runtime/storage/stores/site-config/getters.ts
@@ -1,32 +1,8 @@
 import { computed } from 'vue'
 import type { ComputedRef } from 'vue'
 import type { CwaSiteConfigStateInterface } from './state'
-
-export type SiteConfigParams = {
-  robotsAllowSearchEngineCrawlers: boolean
-  robotsAllowAiBots: boolean
-  robotsText: string
-  robotsRemoveSitemap: boolean
-  sitemapEnabled: boolean
-  siteName: string
-  fallbackTitle: boolean
-  concatTitle: boolean
-  maintenanceModeEnabled: boolean
-  sitemapXml: string
-}
-
-const defaultSettings: SiteConfigParams = {
-  robotsAllowSearchEngineCrawlers: true,
-  robotsAllowAiBots: true,
-  robotsText: '',
-  robotsRemoveSitemap: false,
-  sitemapEnabled: true,
-  siteName: '',
-  fallbackTitle: true,
-  concatTitle: true,
-  maintenanceModeEnabled: false,
-  sitemapXml: '',
-}
+import type { SiteConfigParams } from '#cwa/module'
+import { defaultSiteConfig } from '#cwa/runtime/composables/useCwaSiteConfig'
 
 export interface CwaSiteConfigGettersInterface {
   getConfig: ComputedRef<SiteConfigParams>
@@ -35,7 +11,7 @@ export interface CwaSiteConfigGettersInterface {
 export default function (siteConfigState: CwaSiteConfigStateInterface): CwaSiteConfigGettersInterface {
   return {
     getConfig: computed<SiteConfigParams>(() => {
-      return Object.assign({}, defaultSettings, siteConfigState.config)
+      return Object.assign({}, defaultSiteConfig, siteConfigState.config)
     }),
   }
 }
